requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
deprecation==2.1.0
fastapi==0.135.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
multidict==6.7.1
packaging==26.0
postgrest==2.25.0
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.11.0
pypdf==6.7.4
python-dotenv==1.2.1
realtime==2.25.0
starlette==0.52.1
storage3==2.25.0
StrEnum==0.4.15
supabase==2.25.0
supabase-auth==2.25.0
supabase-functions==2.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
websockets==15.0.1
yarl==1.22.0
